+++ FireworkFlutterSDKSample/lib/models/cart_item.g.dart
@@ -9,14 +9,14 @@ part of 'cart_item.dart';
 CartItem _$CartItemFromJson(Map<String, dynamic> json) => CartItem(
       productId: json['productId'] as String,
       unitId: json['unitId'] as String,
-    )
-      ..title = json['title'] as String?
-      ..subTitle = json['subTitle'] as String?
-      ..amount = json['amount'] as String?
-      ..currencyCode = json['currencyCode'] as String?
-      ..description = json['description'] as String?
-      ..imageURL = json['imageURL'] as String?
-      ..quantity = json['quantity'] as int?;
+      title: json['title'] as String?,
+      subTitle: json['subTitle'] as String?,
+      amount: json['amount'] as String?,
+      currencyCode: json['currencyCode'] as String?,
+      description: json['description'] as String?,
+      imageURL: json['imageURL'] as String?,
+      quantity: json['quantity'] as int?,
+    );
 
 Map<String, dynamic> _$CartItemToJson(CartItem instance) => <String, dynamic>{
       'productId': instance.productId,