+++ CHANGELOG.md
@@ -1,5 +1,75 @@
 # CHANGELOG
 
+## [1.2.0]
+
+### Added
+
+- Support for RTL and Arabic translations
+- Support for autoplay video in thumbnails
+- Add `imageUrl` and `options` to `ProductUnit` class
+- Add `gridColumns` to `VideoFeedConfiguration` class
+- Add `onCustomClickCartIcon` callback in `VideoShopping` class
+
+## [1.2.0-beta.5]
+
+- Firework Flutter SDK beta release
+
+## [1.2.0-beta.4]
+
+- Firework Flutter SDK beta release
+
+## [1.2.0-beta.3]
+
+- Firework Flutter SDK beta release
+
+## [1.2.0-beta.2]
+
+- Firework Flutter SDK beta release
+
+## [1.2.0-beta.1]
+
+- Firework Flutter SDK beta release
+
+## [1.1.3]
+
+### Added
+
+- Add `showBranding` property in `VideoPlayerConfiguration` class
+
+## [1.1.2]
+
+### Added
+
+- Add new properties in `VideoFeedConfiguration` class
+
+## [1.1.1]
+
+### Added
+
+- Add `canPopNativeContainer` method in `FWNavigator` class
+- Add new properties in `FeedItemDetails` class
+
+### Changed
+
+- `popNativeContainer` method also can pop the video or live stream player
+
+## [1.1.0]
+
+### Added
+
+- Add a new video feed source: `dynamicContent`
+- Support for navigating from native page to Flutter page
+
+### Breaking Changes
+
+- Remove `AdConfig` class 
+- Remove `adConfig` parameter from `init` method of `FireworkSDK` class
+- Remove `exitCartPage` method from `VideoShopping` classs
+
+## [1.1.0-beta.1]
+
+- Firework Flutter SDK beta release
+
 ## [1.0.0]
 
 ### Added